componetize
